+++ packages/react/src/index.tsx
@@ -448,4 +448,4 @@ export type { ImageViewerProps } from './components/ImageViewer'
 export { AspectRatio } from './components/AspectRatio'
 export type { AspectRatioProps } from './components/AspectRatio'
 
-export const version = '2.1.1'
+export const version = '2.1.2'